+++ resources/views/layouts/app.blade.php
@@ -16,10 +16,7 @@
 
             <main class="p-4 obiora-main flex-grow-1">
                 @if (session('success'))
-                    <div class="alert alert-success alert-dismissible fade show" role="alert">
-                        {{ session('success') }}
-                        <button type="button" class="btn-close" data-bs-dismiss="alert"></button>
-                    </div>
+                    <span id="obiora-flash-success" class="d-none" data-message="{{ session('success') }}"></span>
                 @endif
 
                 {{ $slot }}